Implemented Valiant's oblivious routing algorithm for Torus.
The class Packet now has a void* for a routing extension. The routing algorithm
is responsible for creating any rounting extensions needed and deleting them.

There was serious bugs in the Torus DimOrder algorithm that made it never know
about crossing datelines. This is bad! Testing shows that it now works right.