Traffic classes and multiple simultaneous applications
This commit implements traffic classes as sets of VCs with corresponding routing
algorithms. The network defines traffic classes and applications are allowed to
use them however they'd like. The applications implemented now have settings
that select which traffic class(es) they use. Traffic classes are set of VCs
that a particular routing algorithm must reside in.

This commit also implements multiple simultaneous applications that run in
parallel. To achieve this there is a new Workload class that manages how
applications run. The settings file specifies an array of applications that will
run. Applications are generally long running (e.g., stresstest and singlestream)
or they are batch style (e.g., simplemem and pulse).

Pulse is a new application that is similar to stresstest except that is doesn't
participate in warm up or cool down. It simply starts up, injects a fixed number
of packets, then shuts off immediately. This is particularly useful for modeling
traffic interference between applications. This can be done with or without
traffic class isolation.