Flexible credit passing to congestion status.
The IQ, IOQ, and OQ router architectures have all been modified such that the
congestion status module can be fed information about available credits and/or
inflight flits in different ways. The IOQ and OQ architectures have 3 modes:
"output", "downstream", and "output_and_downstream". The IQ architecture has 2
modes: "output" and "downstream".

Some restrictions apply to the IQ and OQ
architectures. For IQ, when in "output" mode, the congestion status module
cannot be running in "normalized" AND "vc" mode. This won't work because the IQ
architecture has port-wise output queues so normalizing per VC information is
impossible. For OQ, when in "output" or "output_and_downstream" mode, the
congestion status module must NOT run in normalized mode. How would you
normalize to infinity?

To perform the checks listed above, the congestion status base class now
requires all subclasses to report their 'style' and 'mode'. Router architectures
can use this information to ensure that the congestion status module is
compatible with the architecture.

A destructor check was added to BufferOccupancy to ensure that everything went
back to the init condition. This ensures that credits and/or flits weren't lost
from the BufferOccupancy's perspective.