fixed IOQ OutputAndDownstream mode
The issue is that the output queue has one buffering slot for the switch
allocation phase. This makes one more slot of buffering that wasn't
previously seen. This made the congestion status module believe credits
had been exhausted when they actually weren't. This fix adds 1 more credit
to the congestion status module only in this mode to account for that.